Pocket Calculator with Input from CSV File

+++ src/com/capgemini/calculator/screen/CalculatorScreen.java
@@ -3,7 +3,7 @@ package com.capgemini.calculator.screen;
 public class CalculatorScreen implements Screen{
 
     @Override
-    public void showResult(int result) {
+    public void showResult(double result) {
         System.out.println(result);
     }
 }
