+++ src/com/capgemini/calculator/screen/Screen.java
@@ -1,5 +1,5 @@
 package com.capgemini.calculator.screen;
 
 public interface Screen {
-    void showResult(int result);
+    void showResult(double result);
 }
